refactor: move database dependency to core/dependencies folder to improve readability
